requirements.txt
# requirements.txt

blinker==1.9.0
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
passlib==1.7.4
pymongo==4.12.0
python-dotenv==1.1.0
types-passlib==1.7.7.20250408
Werkzeug==3.1.3
